+++ src/ObjectType/Object/CourseObjectType.php
@@ -2,8 +2,8 @@
 
 namespace srag\Plugins\SrAutoMails\ObjectType\Object;
 
-use ilLPStatusWrapper;
 use ilObjCourse;
+use ilObjUser;
 use srag\Plugins\SrAutoMails\ObjectType\ObjectTypes;
 
 /**
@@ -23,7 +23,11 @@ class CourseObjectType extends ObjObjectType {
 	const OBJECT_PROPERTY_COURSE_END = "course_end";
 	const RECEIVER_COURSE_ADMINISTRATORS = "course_administrators";
 	const RECEIVER_COURSE_MEMBERS = "course_members";
+	const RECEIVER_COURSE_MEMBERS_COMPLETED = "course_members_completed";
+	const RECEIVER_COURSE_MEMBERS_NOT_COMPLETED = "course_members_not_completed";
 	const RECEIVER_COURSE_SUPERIOR_OF_MEMBERS = "course_superior_of_members";
+	const RECEIVER_COURSE_SUPERIOR_OF_MEMBERS_COMPLETED = "course_superior_of_members_completed";
+	const RECEIVER_COURSE_SUPERIOR_OF_MEMBERS_NOT_COMPLETED = "course_superior_of_members_not_completed";
 	const RECEIVER_COURSE_TUTORS = "course_tutors";
 
 
@@ -32,7 +36,34 @@ class CourseObjectType extends ObjObjectType {
 	 * @param array       $placeholders
 	 */
 	protected function applyMailPlaceholders($object, array &$placeholders)/*: void*/ {
+		$members = array_map(function (int $user_id): ilObjUser {
+			return new ilObjUser($user_id);
+		}, $object->getMembersObject()->getMembers());
 
+		$completed = self::ilias()->courses()->getCompletedUsers($object->getId());
+
+		$placeholders = array_merge($placeholders, [
+			"members" => $members,
+			"members_completed" => array_filter($members, function (ilObjUser $user) use ($completed): bool {
+				return in_array($user->getId(), $completed);
+			}),
+			"members_not_completed" => array_filter($members, function (ilObjUser $user) use ($completed): bool {
+				return !in_array($user->getId(), $completed);
+			})
+		]);
+	}
+
+
+	/**
+	 * @inheritdoc
+	 */
+	public function getMailPlaceholderKeyTypes(): array {
+		return array_merge(parent::getMailPlaceholderKeyTypes(), [
+			"object" => "object " . ilObjCourse::class,
+			"members" => "array " . ilObjUser::class,
+			"members_completed" => "array " . ilObjUser::class,
+			"members_not_completed" => "array " . ilObjUser::class,
+		]);
 	}
 
 
@@ -68,7 +99,7 @@ class CourseObjectType extends ObjObjectType {
 				return $object->getCourseEnd();
 
 			case self::OBJECT_PROPERTY_COUNT_COURSE_MEMBERS_COMPLETED:
-				$completed = ilLPStatusWrapper::_lookupCompletedForObject($object->getId());
+				$completed = self::ilias()->courses()->getCompletedUsers($object->getId());
 
 				$completed = array_filter($object->getMembersObject()->getMembers(), function (int $user_id) use ($completed): bool {
 					return in_array($user_id, $completed);
@@ -77,10 +108,10 @@ class CourseObjectType extends ObjObjectType {
 				return count($completed);
 
 			case self::OBJECT_PROPERTY_COUNT_COURSE_MEMBERS_NOT_COMPLETED:
-				$not_completed = ilLPStatusWrapper::_lookupCompletedForObject($object->getId());
+				$completed = self::ilias()->courses()->getCompletedUsers($object->getId());
 
-				$not_completed = array_filter($object->getMembersObject()->getMembers(), function (int $user_id) use ($not_completed): bool {
-					return !in_array($user_id, $not_completed);
+				$not_completed = array_filter($object->getMembersObject()->getMembers(), function (int $user_id) use ($completed): bool {
+					return !in_array($user_id, $completed);
 				});
 
 				return count($not_completed);
@@ -106,6 +137,10 @@ class CourseObjectType extends ObjObjectType {
 		return [
 			self::RECEIVER_COURSE_ADMINISTRATORS => self::RECEIVER_COURSE_ADMINISTRATORS,
 			self::RECEIVER_COURSE_MEMBERS => self::RECEIVER_COURSE_MEMBERS,
+			self::RECEIVER_COURSE_MEMBERS_COMPLETED => self::RECEIVER_COURSE_MEMBERS_COMPLETED,
+			self::RECEIVER_COURSE_MEMBERS_NOT_COMPLETED => self::RECEIVER_COURSE_MEMBERS_NOT_COMPLETED,
+			self::RECEIVER_COURSE_SUPERIOR_OF_MEMBERS_COMPLETED => self::RECEIVER_COURSE_SUPERIOR_OF_MEMBERS_COMPLETED,
+			self::RECEIVER_COURSE_SUPERIOR_OF_MEMBERS_NOT_COMPLETED => self::RECEIVER_COURSE_SUPERIOR_OF_MEMBERS_NOT_COMPLETED,
 			self::RECEIVER_COURSE_SUPERIOR_OF_MEMBERS => self::RECEIVER_COURSE_SUPERIOR_OF_MEMBERS,
 			self::RECEIVER_COURSE_TUTORS => self::RECEIVER_COURSE_TUTORS
 		];
@@ -119,6 +154,8 @@ class CourseObjectType extends ObjObjectType {
 	 * @return int[]
 	 */
 	protected function getReceiverForObject(array $receivers, $object): array {
+		$completed = self::ilias()->courses()->getCompletedUsers($object->getId());
+
 		$array = [];
 
 		foreach ($receivers as $receiver) {
@@ -131,8 +168,36 @@ class CourseObjectType extends ObjObjectType {
 					$array = array_merge($array, $object->getMembersObject()->getMembers());
 					break;
 
+				case self::RECEIVER_COURSE_MEMBERS_COMPLETED:
+					$array = array_merge($array, array_filter($object->getMembersObject()
+						->getMembers(), function (int $user_id) use ($completed): bool {
+						return in_array($user_id, $completed);
+					}));
+					break;
+
+				case self::RECEIVER_COURSE_MEMBERS_NOT_COMPLETED:
+					$array = array_merge($array, array_filter($object->getMembersObject()
+						->getMembers(), function (int $user_id) use ($completed): bool {
+						return !in_array($user_id, $completed);
+					}));
+					break;
+
 				case self::RECEIVER_COURSE_SUPERIOR_OF_MEMBERS:
 					$array = array_merge($array, self::ilias()->orgUnits()->getSuperiorsOfUsers($object->getMembersObject()->getMembers()));
+					break;
+
+				case self::RECEIVER_COURSE_SUPERIOR_OF_MEMBERS_COMPLETED:
+					$array = array_merge($array, self::ilias()->orgUnits()->getSuperiorsOfUsers(array_filter($object->getMembersObject()
+						->getMembers(), function (int $user_id) use ($completed): bool {
+						return in_array($user_id, $completed);
+					})));
+					break;
+
+				case self::RECEIVER_COURSE_SUPERIOR_OF_MEMBERS_NOT_COMPLETED:
+					$array = array_merge($array, self::ilias()->orgUnits()->getSuperiorsOfUsers(array_filter($object->getMembersObject()
+						->getMembers(), function (int $user_id) use ($completed): bool {
+						return !in_array($user_id, $completed);
+					})));
 					break;
 
 				case self::RECEIVER_COURSE_TUTORS: