+++ CHANGELOG.md
@@ -1,5 +1,11 @@
 # Changelog
 
+## [1.3.0]
+- More config interval (Once or number)
+- More config match (Every run or set match)
+- List possible Notifications4Plugins placeholders
+- Expanded superior of members
+
 ## [1.2.0]
 - Supports course members completed and not completed count
 - Each rule can now it's own interval in days to check